+++ libs/ledger-live-common/src/currencies/mock.ts
@@ -482,8 +482,8 @@ export const CURRENCIES_LIST: CryptoCurrency[] = [
     ],
     explorerViews: [
       {
-        address: "https://explorer.near.org/accounts/$address",
-        tx: "https://explorer.near.org/transactions/$hash",
+        address: "https://nearblocks.io/address/$address",
+        tx: "https://nearblocks.io/txns/$hash",
       },
     ],
     keywords: ["near"],