+++ libs/ledger-live-common/src/families/cosmos/js-buildTransaction.test.ts
@@ -12,6 +12,8 @@ import { MsgWithdrawDelegatorReward } from "cosmjs-types/cosmos/distribution/v1b
 import { TxBody, TxRaw } from "cosmjs-types/cosmos/tx/v1beta1/tx";
 import { Fee } from "@keplr-wallet/proto-types/cosmos/tx/v1beta1/tx";
 
+const veryBigNumber = new BigNumber(3333300000000000000000);
+
 describe("txToMessages", () => {
   const transaction: Transaction = {} as Transaction;
   const account: CosmosAccount = {
@@ -34,8 +36,16 @@ describe("txToMessages", () => {
         expect(aminoMsg.type).toContain("MsgSend");
         expect(aminoMsg.value.to_address).toEqual(transaction.recipient);
         expect(aminoMsg.value.from_address).toEqual(account.freshAddress);
-        expect(aminoMsg.value.amount[0].amount).toEqual(transaction.amount.toString());
+        expect(aminoMsg.value.amount[0].amount).toEqual(transaction.amount.toFixed());
         expect(aminoMsg.value.amount[0].denom).toEqual(account.currency.units[1].code);
+      });
+
+      it("should not include exponential part on big numbers", () => {
+        transaction.recipient = "address";
+        transaction.amount = veryBigNumber;
+        const { aminoMsgs } = txToMessages(account, transaction);
+        const [aminoMsg] = aminoMsgs;
+        expect(aminoMsg.value.amount[0].amount.includes("e")).toEqual(false);
       });
 
       it("should return no message if recipient isn't defined", () => {
@@ -75,8 +85,17 @@ describe("txToMessages", () => {
         expect(protoMsg.typeUrl).toContain("MsgSend");
         expect(value.toAddress).toEqual(transaction.recipient);
         expect(value.fromAddress).toEqual(account.freshAddress);
-        expect(value.amount[0].amount).toEqual(transaction.amount.toString());
+        expect(value.amount[0].amount).toEqual(transaction.amount.toFixed());
         expect(value.amount[0].denom).toEqual(account.currency.units[1].code);
+      });
+
+      it("should not include exponential part on big numbers", () => {
+        transaction.recipient = "address";
+        transaction.amount = veryBigNumber;
+        const { protoMsgs } = txToMessages(account, transaction);
+        const [protoMsg] = protoMsgs;
+        const value = cosmos.bank.v1beta1.MsgSend.decode(protoMsg.value);
+        expect(value.amount[0].amount?.includes("e")).toEqual(false);
       });
 
       it("should return no message if recipient isn't defined", () => {
@@ -120,8 +139,22 @@ describe("txToMessages", () => {
         expect(message.type).toContain("MsgDelegate");
         expect(message.value.validator_address).toEqual(transaction.validators[0].address);
         expect(message.value.delegator_address).toEqual(account.freshAddress);
-        expect(message.value.amount?.amount).toEqual(transaction.amount.toString());
+        expect(message.value.amount?.amount).toEqual(transaction.amount.toFixed());
         expect(message.value.amount?.denom).toEqual(account.currency.units[1].code);
+      });
+
+      it("should not include exponential part on big numbers", () => {
+        transaction.recipient = "address";
+        transaction.amount = veryBigNumber;
+        transaction.validators = [
+          {
+            address: "realAddressTrustMe",
+            amount: veryBigNumber,
+          } as CosmosDelegationInfo,
+        ];
+        const { aminoMsgs } = txToMessages(account, transaction);
+        const [message] = aminoMsgs;
+        expect(message.value.amount?.amount.includes("e")).toEqual(false);
       });
 
       it("should return no message if tx has a 0 amount", () => {
@@ -167,8 +200,22 @@ describe("txToMessages", () => {
         const value = MsgDelegate.decode(message.value);
         expect(value.validatorAddress).toEqual(transaction.validators[0].address);
         expect(value.delegatorAddress).toEqual(account.freshAddress);
-        expect(value.amount?.amount).toEqual(transaction.amount.toString());
+        expect(value.amount?.amount).toEqual(transaction.amount.toFixed());
         expect(value.amount?.denom).toEqual(account.currency.units[1].code);
+      });
+
+      it("should not include exponential part on big numbers", () => {
+        transaction.amount = veryBigNumber;
+        transaction.validators = [
+          {
+            address: "realAddressTrustMe",
+            amount: veryBigNumber,
+          } as CosmosDelegationInfo,
+        ];
+        const { protoMsgs } = txToMessages(account, transaction);
+        const [message] = protoMsgs;
+        const value = MsgDelegate.decode(message.value);
+        expect(value.amount?.amount.includes("e")).toEqual(false);
       });
 
       it("should return no message if tx has a 0 amount", () => {
@@ -219,8 +266,21 @@ describe("txToMessages", () => {
         expect(message.type).toContain("MsgUndelegate");
         expect(message.value.validator_address).toEqual(transaction.validators[0].address);
         expect(message.value.delegator_address).toEqual(account.freshAddress);
-        expect(message.value.amount?.amount).toEqual(transaction.validators[0].amount.toString());
+        expect(message.value.amount?.amount).toEqual(transaction.validators[0].amount.toFixed());
         expect(message.value.amount?.denom).toEqual(account.currency.units[1].code);
+      });
+
+      it("should not include exponential part on big numbers", () => {
+        transaction.amount = veryBigNumber;
+        transaction.validators = [
+          {
+            address: "realAddressTrustMe",
+            amount: veryBigNumber,
+          } as CosmosDelegationInfo,
+        ];
+        const { aminoMsgs } = txToMessages(account, transaction);
+        const [message] = aminoMsgs;
+        expect(message.value.amount?.amount.includes("e")).toEqual(false);
       });
 
       it("should return no message if validators aren't defined", () => {
@@ -279,8 +339,22 @@ describe("txToMessages", () => {
         const value = MsgUndelegate.decode(message.value);
         expect(value.validatorAddress).toEqual(transaction.validators[0].address);
         expect(value.delegatorAddress).toEqual(account.freshAddress);
-        expect(value.amount?.amount).toEqual(transaction.validators[0].amount.toString());
+        expect(value.amount?.amount).toEqual(transaction.validators[0].amount.toFixed());
         expect(value.amount?.denom).toEqual(account.currency.units[1].code);
+      });
+
+      it("should not include exponential part on big numbers", () => {
+        transaction.amount = veryBigNumber;
+        transaction.validators = [
+          {
+            address: "realAddressTrustMe",
+            amount: veryBigNumber,
+          } as CosmosDelegationInfo,
+        ];
+        const { protoMsgs } = txToMessages(account, transaction);
+        const [message] = protoMsgs;
+        const value = MsgUndelegate.decode(message.value);
+        expect(value.amount?.amount.includes("e")).toEqual(false);
       });
 
       it("should return no message if validators aren't defined", () => {
@@ -345,8 +419,21 @@ describe("txToMessages", () => {
         expect(message.value.validator_src_address).toEqual(transaction.sourceValidator);
         expect(message.value.validator_dst_address).toEqual(transaction.validators[0].address);
         expect(message.value.delegator_address).toEqual(account.freshAddress);
-        expect(message.value.amount.amount).toEqual(transaction.validators[0].amount.toString());
+        expect(message.value.amount.amount).toEqual(transaction.validators[0].amount.toFixed());
         expect(message.value.amount.denom).toEqual(account.currency.units[1].code);
+      });
+
+      it("should not include exponential part on big numbers", () => {
+        transaction.sourceValidator = "source";
+        transaction.validators = [
+          {
+            address: "realAddressTrustMe",
+            amount: veryBigNumber,
+          } as CosmosDelegationInfo,
+        ];
+        const { aminoMsgs } = txToMessages(account, transaction);
+        const [message] = aminoMsgs;
+        expect(message.value.amount.amount.includes("e")).toEqual(false);
       });
 
       it("should return no message if sourceValidator isn't defined", () => {
@@ -410,8 +497,23 @@ describe("txToMessages", () => {
         expect(value.validatorSrcAddress).toEqual(transaction.sourceValidator);
         expect(value.validatorDstAddress).toEqual(transaction.validators[0].address);
         expect(value.delegatorAddress).toEqual(account.freshAddress);
-        expect(value.amount?.amount).toEqual(transaction.validators[0].amount.toString());
+        expect(value.amount?.amount).toEqual(transaction.validators[0].amount.toFixed());
         expect(value.amount?.denom).toEqual(account.currency.units[1].code);
+      });
+
+      it("should not include exponential part on big numbers", () => {
+        transaction.sourceValidator = "source";
+        transaction.validators = [
+          {
+            address: "realAddressTrustMe",
+            amount: veryBigNumber,
+          } as CosmosDelegationInfo,
+        ];
+        const { protoMsgs } = txToMessages(account, transaction);
+        const [message] = protoMsgs;
+        expect(message).toBeTruthy();
+        const value = MsgBeginRedelegate.decode(message.value);
+        expect(value.amount?.amount.includes("e")).toEqual(false);
       });
 
       it("should return no message if sourceValidator isn't defined", () => {
@@ -558,9 +660,21 @@ describe("txToMessages", () => {
         expect(delegateMessage.value.validator_address).toEqual(transaction.validators[0].address);
         expect(delegateMessage.value.delegator_address).toEqual(account.freshAddress);
         expect(delegateMessage.value.amount.amount).toEqual(
-          transaction.validators[0].amount.toString(),
+          transaction.validators[0].amount.toFixed(),
         );
         expect(delegateMessage.value.amount.denom).toEqual(account.currency.units[1].code);
+      });
+
+      it("should not include exponential part on big numbers", () => {
+        transaction.validators = [
+          {
+            address: "iAmAValidatorAddress",
+            amount: veryBigNumber,
+          } as CosmosDelegationInfo,
+        ];
+        const { aminoMsgs } = txToMessages(account, transaction);
+        const [, delegateMessage] = aminoMsgs;
+        expect(delegateMessage.value.amount.amount.includes("e")).toEqual(false);
       });
     });
 
@@ -585,9 +699,22 @@ describe("txToMessages", () => {
         expect(delegateMessageValue.validatorAddress).toEqual(transaction.validators[0].address);
         expect(delegateMessageValue.delegatorAddress).toEqual(account.freshAddress);
         expect(delegateMessageValue.amount?.amount).toEqual(
-          transaction.validators[0].amount.toString(),
+          transaction.validators[0].amount.toFixed(),
         );
         expect(delegateMessageValue.amount?.denom).toEqual(account.currency.units[1].code);
+      });
+
+      it("should return a MsgWithdrawDelegatorReward message and a MsgDelegate if transaction is complete", () => {
+        transaction.validators = [
+          {
+            address: "iAmAValidatorAddress",
+            amount: veryBigNumber,
+          } as CosmosDelegationInfo,
+        ];
+        const { protoMsgs } = txToMessages(account, transaction);
+        const [, delegateMessage] = protoMsgs;
+        const delegateMessageValue = MsgDelegate.decode(delegateMessage.value);
+        expect(delegateMessageValue.amount?.amount.includes("e")).toEqual(false);
       });
     });
   });